resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.Restaurants.Commands.CreateRestaurant;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes
@@ -11,12 +13,14 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
         private readonly ILogger<CreateRestaurantCommandHandler> _logger;
         private readonly IRestaurantsRepository _restaurantsRepository;
         private readonly IDishesRepository _dishesRepository;
+        private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
 
-        public DeleteDishesForRestaurantCommandHandler(ILogger<CreateRestaurantCommandHandler> logger, IRestaurantsRepository restaurantsRepository, IDishesRepository dishesRepository)
+        public DeleteDishesForRestaurantCommandHandler(ILogger<CreateRestaurantCommandHandler> logger, IRestaurantsRepository restaurantsRepository, IDishesRepository dishesRepository, IRestaurantAuthorizationService restaurantAuthorizationService)
         {
             _logger = logger;
             _restaurantsRepository = restaurantsRepository;
             _dishesRepository = dishesRepository;
+            _restaurantAuthorizationService = restaurantAuthorizationService;
         }
         public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -25,6 +29,9 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
             var restaurant = await _restaurantsRepository.GetAsync(request.RestaurantId);
             if (restaurant == null)
                 throw new NotFoundException(nameof(restaurant), request.RestaurantId.ToString());
+
+            if (!_restaurantAuthorizationService.Authorize(restaurant, ResourceEnum.Update))
+                throw new ForbidException();
 
             await _dishesRepository.Delete(restaurant.Dishes);
         }
